requirements.txt
django
djangorestframework
python-dotenv

sentence-transformers
torch
transformers

langchain
langchain-community
langchain-groq

pinecone
pypdf
djangorestframework-simplejwt